% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLcompare-package.R
\docType{package}
\name{MLcompare-package}
\alias{MLcompare}
\alias{MLcompare-package}
\title{MLcompare: Compare Machine Learning Algorithms Performance}
\description{
Compare the performance of deep learning (implemented in H2O) and six other machine learning algorithms implemented in caret for classification tasks.
}
\author{
\strong{Maintainer}: Dany Mukesha \email{danymukesha@gmail.com} (\href{https://orcid.org/0009-0001-9514-751X}{ORCID})

}
\keyword{internal}
