% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocessing.R
\name{preprocess_data}
\alias{preprocess_data}
\title{Preprocess Data}
\usage{
preprocess_data(data)
}
\arguments{
\item{data}{Data frame containing the metabolite data and subtype labels.}
}
\value{
A data frame with quantile-normalized metabolite data.
}
\description{
Preprocess Data
}
