% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_testing.R
\name{test_models}
\alias{test_models}
\title{Test Models}
\usage{
test_models(models, test_data)
}
\arguments{
\item{models}{A list of trained models.}

\item{test_data}{Testing data frame.}
}
\value{
A data frame with performance metrics for each model.
}
\description{
Test Models
}
