% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_training.R
\name{train_deep_learning_model}
\alias{train_deep_learning_model}
\title{Train Deep Learning Model using H2O}
\usage{
train_deep_learning_model(
  data,
  hidden_layers = c(50, 50),
  epochs = 10,
  activation = "RectifierWithDropout"
)
}
\arguments{
\item{data}{Training data frame with predictors and response variable.}

\item{hidden_layers}{A vector specifying the number of neurons
in each hidden layer (e.g., c(50, 50)).}

\item{epochs}{Number of epochs for training.}

\item{activation}{Activation function to use, e.g., 'RectifierWithDropout'.}
}
\value{
A trained H2O deep learning model.
}
\description{
Train Deep Learning Model using H2O
}
