% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_training.R
\name{train_models}
\alias{train_models}
\title{Train Basic ML Models}
\usage{
train_models(data, algorithms = c("rpart", "lda", "svmRadial", "rf", "gbm"))
}
\arguments{
\item{data}{Training data frame.}

\item{algorithms}{A vector of algorithms to train,
such as c('rpart', 'lda', 'svmRadial', 'rf', 'gbm').}
}
\value{
A list of trained models.
}
\description{
Train Basic ML Models
}
